socialapi/countly: add group cache
